Changes to image loader to prevent images from being cached forever.
When getting an image from the cache it will check if the image has expired.
if it's expired, it will remove the image from the cache and act as though it was never there.

This should fix issue #10 whereby the images were not being updated until the cache is cleared.