It would appear sometimes the status' include more than the last ones.
Since there is one which is new, the original if rules are OK, but the counts are way out.

Added a filtering to ensure all of the statuses are new and then the rest of the logic runs on the filtered list.

This really should fix #110 (i managed to replicate this whist a debugger was running)